create to-do ui
- move to-do ui to right file
- also bring nav, side bar to to-do page
